+++ src/main/scala/contract/callback/todo/ItemCallback.scala
@@ -5,12 +5,12 @@ import scala.concurrent.Future
 
 abstract class ItemCallback {
 
-  def add(userId: Int, body: String, state: Boolean): Future[Map[Int, Item]]
+  def add(userId: Long, body: String, state: Boolean): Future[Map[Long, Item]]
 
-  def getById(userId: Int): Future[Option[Map[Int, Item]]]
+  def getById(userId: Long): Future[Option[Map[Long, Item]]]
 
-  def editBody(userId: Int, id: Int, newBody: String): Future[Map[Int, Item]]
+  def editBody(userId: Long, id: Long, newBody: String): Future[Map[Long, Item]]
 
-  def editState(userId: Int, id: Int, newState: Boolean): Future[Map[Int, Item]]
+  def editState(userId: Long, id: Long, newState: Boolean): Future[Map[Long, Item]]
 
 }
