+++ src/main/scala/contract/service/auth/SignOutService.scala
@@ -6,6 +6,6 @@ abstract class SignOutService extends Service[SignOutService.Request, Unit]
 
 object SignOutService {
 
-  case class Request(userID: Int)
+  case class Request(userID: Long)
 
 }
