+++ src/main/scala/domain/auth/User.scala
@@ -1,6 +1,6 @@
 package domain.auth
 
-case class User(id: Int, username: String, password: String) {
+case class User(id: Long, username: String, password: String) {
 
   def setUsername(newUsername: String): User = {
     copy(username = newUsername)
